start moving routes into subpackage
- deleted lots of old code
- moved fishy route into routes (almost finished)
- moved all other routes into routes/todo.go (for later)
- replaced main.go with fasthttp server and viper config loader
